REFACTOR: 멤버 관련된 집안일 API 패키지 변경
